# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/ai/config:test:
        post:
            tags:
                - AIService
            description: TestAIConfig tests the AI configuration by sending a test request to the AI provider.
            operationId: AIService_TestAIConfig
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TestAIConfigRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TestAIConfigResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai/summaries:generate:
        post:
            tags:
                - AIService
            description: GenerateAISummary generates an AI summary for memos within a specified time range and tags.
            operationId: AIService_GenerateAISummary
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenerateAISummaryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Memo'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        Attachment:
            required:
                - filename
                - type
            type: object
            properties:
                name:
                    type: string
                    description: "The name of the attachment.\r\n Format: attachments/{attachment}"
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
                filename:
                    type: string
                    description: The filename of the attachment.
                content:
                    writeOnly: true
                    type: string
                    description: Input only. The content of the attachment.
                    format: bytes
                externalLink:
                    type: string
                    description: Optional. The external link of the attachment.
                type:
                    type: string
                    description: The MIME type of the attachment.
                size:
                    readOnly: true
                    type: string
                    description: Output only. The size of the attachment in bytes.
                memo:
                    type: string
                    description: "Optional. The related memo. Refer to `Memo.name`.\r\n Format: memos/{memo}"
...
        BlockquoteNode:
            type: object
            properties:
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
...
        CodeBlockNode:
            type: object
            properties:
                language:
                    type: string
                content:
                    type: string
...
        GenerateAISummaryRequest:
            required:
                - timeRange
            type: object
            properties:
                timeRange:
                    type: string
                    description: "The time range for selecting source memos.\r\n Supported values: \"7d\", \"30d\", \"90d\", \"custom\"\r\n If \"custom\" is specified, start_date and end_date must be provided."
                tags:
                    type: array
                    items:
                        type: string
                    description: "Optional. Tags to filter source memos.\r\n Only memos containing these tags will be included in the summary."
                startDate:
                    type: string
                    description: "Optional. The start date for custom time range.\r\n Format: YYYY-MM-DD\r\n Required when time_range is \"custom\"."
                endDate:
                    type: string
                    description: "Optional. The end date for custom time range.\r\n Format: YYYY-MM-DD\r\n Required when time_range is \"custom\"."
            description: Request message for GenerateAISummary method.
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        HeadingNode:
            type: object
            properties:
                level:
                    type: integer
                    format: int32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
...
        HorizontalRuleNode:
            type: object
            properties:
                symbol:
                    type: string
...
        LineBreakNode:
            type: object
            properties: {}
...
        ListNode:
            type: object
            properties:
                kind:
                    enum:
                        - KIND_UNSPECIFIED
                        - ORDERED
                        - UNORDERED
                        - DESCRIPTION
                    type: string
                    format: enum
                indent:
                    type: integer
                    format: int32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
...
        Location:
            type: object
            properties:
                placeholder:
                    type: string
                    description: A placeholder text for the location.
                latitude:
                    type: number
                    description: The latitude of the location.
                    format: double
                longitude:
                    type: number
                    description: The longitude of the location.
                    format: double
        MathBlockNode:
            type: object
            properties:
                content:
                    type: string
...
        Memo:
            required:
                - state
                - content
                - visibility
            type: object
            properties:
                name:
                    type: string
                    description: "The resource name of the memo.\r\n Format: memos/{memo}, memo is the user defined id or uuid."
                state:
                    enum:
                        - STATE_UNSPECIFIED
                        - NORMAL
                        - ARCHIVED
                    type: string
                    description: The state of the memo.
                    format: enum
                creator:
                    readOnly: true
                    type: string
                    description: "The name of the creator.\r\n Format: users/{user}"
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    description: Output only. The last update timestamp.
                    format: date-time
                displayTime:
                    type: string
                    description: The display timestamp of the memo.
                    format: date-time
                content:
                    type: string
                    description: Required. The content of the memo in Markdown format.
                nodes:
                    readOnly: true
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
                    description: Output only. The parsed nodes from the content.
                visibility:
                    enum:
                        - VISIBILITY_UNSPECIFIED
                        - PRIVATE
                        - PROTECTED
                        - PUBLIC
                    type: string
                    description: The visibility of the memo.
                    format: enum
                tags:
                    readOnly: true
                    type: array
                    items:
                        type: string
                    description: Output only. The tags extracted from the content.
                pinned:
                    type: boolean
                    description: Whether the memo is pinned.
                attachments:
                    type: array
                    items:
                        $ref: '#/components/schemas/Attachment'
                    description: Optional. The attachments of the memo.
                relations:
                    type: array
                    items:
                        $ref: '#/components/schemas/MemoRelation'
                    description: Optional. The relations of the memo.
                reactions:
                    readOnly: true
                    type: array
                    items:
                        $ref: '#/components/schemas/Reaction'
                    description: Output only. The reactions to the memo.
                property:
                    readOnly: true
                    allOf:
                        - $ref: '#/components/schemas/Memo_Property'
                    description: Output only. The computed properties of the memo.
                parent:
                    readOnly: true
                    type: string
                    description: "Output only. The name of the parent memo.\r\n Format: memos/{memo}"
                snippet:
                    readOnly: true
                    type: string
                    description: Output only. The snippet of the memo content. Plain text only.
                location:
                    allOf:
                        - $ref: '#/components/schemas/Location'
                    description: Optional. The location of the memo.
        MemoRelation:
            required:
                - memo
                - relatedMemo
                - type
            type: object
            properties:
                memo:
                    allOf:
                        - $ref: '#/components/schemas/MemoRelation_Memo'
                    description: The memo in the relation.
                relatedMemo:
                    allOf:
                        - $ref: '#/components/schemas/MemoRelation_Memo'
                    description: The related memo.
                type:
                    enum:
                        - TYPE_UNSPECIFIED
                        - REFERENCE
                        - COMMENT
                    type: string
                    format: enum
...
        Memo_Property:
            type: object
            properties:
                hasLink:
                    type: boolean
                hasTaskList:
                    type: boolean
                hasCode:
                    type: boolean
                hasIncompleteTasks:
                    type: boolean
            description: Computed properties of a memo.
        Node:
            type: object
            properties:
                type:
                    enum:
                        - NODE_UNSPECIFIED
                        - LINE_BREAK
                        - PARAGRAPH
                        - CODE_BLOCK
                        - HEADING
                        - HORIZONTAL_RULE
                        - BLOCKQUOTE
                        - LIST
                        - ORDERED_LIST_ITEM
                        - UNORDERED_LIST_ITEM
                        - TASK_LIST_ITEM
                        - MATH_BLOCK
                        - TABLE
                        - EMBEDDED_CONTENT
                        - TEXT
                        - BOLD
                        - ITALIC
                        - BOLD_ITALIC
                        - CODE
                        - IMAGE
                        - LINK
                        - AUTO_LINK
                        - TAG
                        - STRIKETHROUGH
                        - ESCAPING_CHARACTER
                        - MATH
                        - HIGHLIGHT
                        - SUBSCRIPT
                        - SUPERSCRIPT
                        - REFERENCED_CONTENT
                        - SPOILER
                        - HTML_ELEMENT
                    type: string
                    format: enum
                lineBreakNode:
                    allOf:
                        - $ref: '#/components/schemas/LineBreakNode'
                    description: Block nodes.
                paragraphNode:
                    $ref: '#/components/schemas/ParagraphNode'
                codeBlockNode:
                    $ref: '#/components/schemas/CodeBlockNode'
                headingNode:
                    $ref: '#/components/schemas/HeadingNode'
                horizontalRuleNode:
                    $ref: '#/components/schemas/HorizontalRuleNode'
                blockquoteNode:
                    $ref: '#/components/schemas/BlockquoteNode'
                listNode:
                    $ref: '#/components/schemas/ListNode'
                orderedListItemNode:
                    $ref: '#/components/schemas/OrderedListItemNode'
                unorderedListItemNode:
                    $ref: '#/components/schemas/UnorderedListItemNode'
                taskListItemNode:
                    $ref: '#/components/schemas/TaskListItemNode'
                mathBlockNode:
                    $ref: '#/components/schemas/MathBlockNode'
                tableNode:
                    $ref: '#/components/schemas/TableNode'
                embeddedContentNode:
                    $ref: '#/components/schemas/EmbeddedContentNode'
                textNode:
                    allOf:
                        - $ref: '#/components/schemas/TextNode'
                    description: Inline nodes.
                boldNode:
                    $ref: '#/components/schemas/BoldNode'
                italicNode:
                    $ref: '#/components/schemas/ItalicNode'
                boldItalicNode:
                    $ref: '#/components/schemas/BoldItalicNode'
                codeNode:
                    $ref: '#/components/schemas/CodeNode'
                imageNode:
                    $ref: '#/components/schemas/ImageNode'
                linkNode:
                    $ref: '#/components/schemas/LinkNode'
                autoLinkNode:
                    $ref: '#/components/schemas/AutoLinkNode'
                tagNode:
                    $ref: '#/components/schemas/TagNode'
                strikethroughNode:
                    $ref: '#/components/schemas/StrikethroughNode'
                escapingCharacterNode:
                    $ref: '#/components/schemas/EscapingCharacterNode'
                mathNode:
                    $ref: '#/components/schemas/MathNode'
                highlightNode:
                    $ref: '#/components/schemas/HighlightNode'
                subscriptNode:
                    $ref: '#/components/schemas/SubscriptNode'
                superscriptNode:
                    $ref: '#/components/schemas/SuperscriptNode'
                referencedContentNode:
                    $ref: '#/components/schemas/ReferencedContentNode'
                spoilerNode:
                    $ref: '#/components/schemas/SpoilerNode'
                htmlElementNode:
                    $ref: '#/components/schemas/HTMLElementNode'
...
        OrderedListItemNode:
            type: object
            properties:
                number:
                    type: string
                indent:
                    type: integer
                    format: int32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
        ParagraphNode:
            type: object
            properties:
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Node'
...
        Reaction:
            required:
                - contentId
                - reactionType
            type: object
            properties:
                name:
                    readOnly: true
                    type: string
                    description: "The resource name of the reaction.\r\n Format: reactions/{reaction}"
                creator:
                    readOnly: true
                    type: string
                    description: "The resource name of the creator.\r\n Format: users/{user}"
                contentId:
                    type: string
                    description: "The resource name of the content.\r\n For memo reactions, this should be the memo's resource name.\r\n Format: memos/{memo}"
                reactionType:
                    type: string
                    description: "Required. The type of reaction (e.g., \"\U0001F44D\", \"❤️\", \"\U0001F604\")."
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TestAIConfigRequest:
            type: object
            properties: {}
            description: Request message for TestAIConfig method.
        TestAIConfigResponse:
            type: object
            properties:
                success:
                    type: boolean
                    description: Whether the AI configuration test was successful.
                errorMessage:
                    type: string
                    description: Optional. Error message if the test failed.
                details:
                    type: string
                    description: Optional. Additional details about the test result.
            description: Response message for TestAIConfig method.
...